new file structure
This removes the `public` directory entirely. Instead, images are inside
`src/images` and `index.html` is inside `src` as well. This removes the
requirement for the copy plugin in favor for the Webpack file loader.
This was done to keep a consistent structure where everything inside
`src` forms a module, and this module will have access to images usable
by any submodule. It also switches to absolute paths by default,
treating everything inside `src` as a module.